refactor: remove `@/components/bsx`
