--- components/bsx/Create/utils.ts
@@ -1,7 +1,0 @@
-export type Token = 'DOT' | 'KSM'
-
-export const getBalance = (_token?: Token): number => {
-  const { balance } = useAuth()
-
-  return Number(balance.value)
-}
